% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_data_type.R
\name{assign_data_type}
\alias{assign_data_type}
\title{Assign a type of data given a set of data types to match.}
\usage{
assign_data_type(x, col_names, no_match = warning)
}
\arguments{
\item{x}{A data.frame or list of data.frames to detect the data type}

\item{col_names}{A named list where the names are the data type and the
values are required column names (additional column names may be present in
the data).}

\item{no_match}{What should be done if no data type matches `x`?  (Typically
this will be one of `stop`, `warning`, or `message`, but it can be any
function taking a single character argument.)}
}
\value{
`x` with classes of the data type and "filetype_mappings" added.
}
\description{
Assign a type of data given a set of data types to match.
}
\details{
To detect a data type, first `col_names` is sorted by number of
  columns provided so that more matching names are matched rather than fewer.
  Then, the first among the sorted `col_names` list is selected.
}
