% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftm_clean.R
\name{ftm_clean}
\alias{ftm_clean}
\title{Clean data of a given class}
\usage{
ftm_clean(data, ...)
}
\arguments{
\item{data}{The data to clean}

\item{...}{options for cleaning passed to methods}
}
\value{
The cleaned data
}
\description{
To use this, you must write a method that is specific to that data class.
Typically, data will have been loaded, run through `assign_data_type()` to
have the class assigned, and then run through `ftm_clean()` to clean any
idiosyncracies of how data may not be clean from the source.  `ftm_clean()`
is not typically used to change column names or make other modifications.
}
