% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_data_type.R
\name{register_data_type}
\alias{register_data_type}
\alias{register_data_type_instance}
\title{Generate a function to store and retrieve data column names for use with
assign_data_type}
\usage{
register_data_type()

register_data_type_instance(.l = list(), ..., overwrite = FALSE, reset = FALSE)
}
\arguments{
\item{.l}{A a named list of data types.  Names are the name of the data
source and values are column names.}

\item{...}{Named arguments interpreted the same as `.l`}
}
\value{
A function to store and retrieve data column names
}
\description{
Generate a function to store and retrieve data column names for use with
assign_data_type

A specific example of how to use `register_data_type()`.  Do not use
`register_data_type_instance()` in your own package; generate your own
instance as shown in the example.
}
\examples{
my_data_types <- register_data_type()
my_data_types(data_source_1=c("A", "B"))
my_data_types()
}
